Support for C-Style functions
~ moved all of the main functionality to glfw/GLFW.py
~ using mixedCase in glfw/GLFW.py and
~ using snake_case in glfw/__init__.py

+ added leading '_' to imported modules in glfw/GLFW.py (to exclude them from the snake_case generator)
+ added _error_reporting_query_func to glfw/GLFW.py to pass through the ERROR_REPORTING macro
